created "Track" class that extends Sound, pulled some functionality out of Sound
