+++ addon/utils/sound.js
@@ -1,6 +1,6 @@
 import Ember from 'ember';
 
-export const Sound = Ember.Object.extend({
+const Sound = Ember.Object.extend({
   name: null,
   node: null,
   panner: null,
@@ -15,57 +15,6 @@ export const Sound = Ember.Object.extend({
   analyzerNode: null,
   analyzePreGain: false,
 
-  _initNodes: Ember.on('init', function() {
-    const ctx = this.get('audioContext');
-
-    this.set('pannerNode', ctx.createStereoPanner());
-    this.set('gainNode', ctx.createGain());
-    this.set('analyzerNode', ctx.createAnalyser());
-  }),
-
-  position: Ember.computed('startOffset', function() {
-    const startOffset = this.get('startOffset');
-    let minutes = Math.floor(startOffset / 60);
-    let seconds = (startOffset - (minutes * 60)).toFixed();
-
-    if (seconds === '60') {
-      seconds = '00';
-      minutes += 1;
-    } else if (seconds < 10) {
-      seconds = `0${seconds}`;
-    }
-
-    return {
-      raw: startOffset,
-      string: `${minutes}:${seconds}`,
-      obj: { minutes, seconds }
-    };
-  }),
-
-  percentPlayed: Ember.computed('duration', 'startOffset', function() {
-    const ratio = this.get('startOffset') / this.get('duration.raw');
-    return ratio * 100;
-  }),
-
-  percentGain: Ember.computed('gainNode.gain.value', function() {
-    return this.get('gainNode.gain.value') * 100;
-  }),
-
-  watchPosition: Ember.observer('isPlaying', function() {
-    const ctx = this.get('audioContext');
-    const startOffset = this.get('startOffset');
-    const startTime = this.get('startTime');
-
-    const animate = () => {
-      if (this.get('isPlaying')) {
-        this.set('startOffset', startOffset + ctx.currentTime - startTime);
-        requestAnimationFrame(animate);
-      }
-    };
-
-    requestAnimationFrame(animate);
-  }),
-
   duration: Ember.computed('audioBuffer.duration', function() {
     const duration = this.get('audioBuffer.duration');
     const minutes = (duration / 60).toFixed();
@@ -76,6 +25,18 @@ export const Sound = Ember.Object.extend({
       string: `${minutes}:${seconds}`,
       obj: { minutes, seconds }
     };
+  }),
+
+  percentGain: Ember.computed('gainNode.gain.value', function() {
+    return this.get('gainNode.gain.value') * 100;
+  }),
+
+  _initNodes: Ember.on('init', function() {
+    const ctx = this.get('audioContext');
+
+    this.set('pannerNode', ctx.createStereoPanner());
+    this.set('gainNode', ctx.createGain());
+    this.set('analyzerNode', ctx.createAnalyser());
   }),
 
   /**
@@ -178,3 +139,5 @@ export const Sound = Ember.Object.extend({
     };
   }
 });
+
+export default Sound;
