+++ tests/unit/utils/track-test.js
@@ -1,0 +1,10 @@
+import track from 'dummy/utils/track';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | track');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  let result = track();
+  assert.ok(result);
+});
